requirements.txt
Django~=2.2.10
djangorestframework
